backend/python_services/requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Time Series Forecasting
prophet==1.1.5
statsmodels==0.14.1
pmdarima==2.0.4

# Database Connectivity
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
pymysql==1.1.0

# API and Web Framework
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
joblib==1.3.2
